refactor: Attribute 타입은 types 폴더로 이동
