+++ types/index.ts
@@ -2,3 +2,8 @@ export type UnPack<T> = T extends (infer U)[] ? U : T;
 export type PropsWithChildren<P = unknown> = P & {
   children?: string;
 };
+
+export interface Attribute {
+  class: string;
+  [key: string]: string;
+}
